# Speed up Ghost boot time