--- a/package.json
+++ b/package.json
@@ -89,7 +89,7 @@
     "js-yaml": "3.12.1",
     "jsonpath": "1.0.0",
     "jsonwebtoken": "8.4.0",
-    "keypair": "1.0.1",
+    "rsa-keypair": "1.0.1",
     "knex": "0.14.6",
     "knex-migrator": "3.2.5",
     "lodash": "4.17.11",