Speed up Ghost boot time by replacing the JS-only RSA key generator library with a native alternative.

---

After profiling the initial boot, I saw the key generation library took around 25% of the boot time. (Flat section in the middle with red at the top.) JS-only libraries tend to be pretty slow for crypto.

![image](https://user-images.githubusercontent.com/964245/58364562-59c7df80-7ef1-11e9-8562-9715aea9768b.png)

I benchmarked some alternatives and [`rsa-keypair`](https://github.com/sunjith/node-rsa-keypair) (a native lib) is much faster.

```
keypair x 3.99 ops/sec ±34.45% (15 runs sampled)
rsa-keypair x 48.53 ops/sec ±8.31% (23 runs sampled)
Fastest is rsa-keypair
```

Server (in development mode) boot time on my laptop:
* Before: ~2.1s
* After: ~1.7s

So it looks like it's about 20% faster for me.

Just required a quick switch-out but should work the same. :+1: