@kevinansfield 2019-05-28
Hey @neosilky 👋 Thanks for this, looks really interesting! In general we've tried to keep away from native/binary dependencies because they've historically caused a lot of install problems for users across the myriad of environments in which Ghost is used. That said, startup performance is important to us so I've raised this internally to review in more detail 🙂