@kevinansfield 2019-05-28
Looking at where `keypair` is used it looks like there may be a potential refactoring to skip it altogether on boot.

The `dynamicDefault` object is used to populate default settings values for when they don't already exist, however it's still generated on every boot. Normally this isn't much of an issue because it's really quick but the usage of `keypair` has changed that. It may be that we can refactor the members keypair generation so that it only happens once on first boot, or we can move the generation to where it's used and only run it if the keypair is not already stored in settings. /cc @allouis, @rishabhgrg