@neosilky 2019-05-28
@kevinansfield Sure thing! The library only relies on OpenSSL but maybe there are situations where this could cause a problem.

I do have some other changes that would help performance. They're just JavaScript changes but mostly requires switching pretty lodash code for uglier vanilla code. Probably not something you want to do but it's an option.