@kevinansfield 2019-05-28
> I do have some other changes that would help performance. They're just JavaScript changes but mostly requires switching pretty lodash code for uglier vanilla code. Probably not something you want to do but it's an option.

If they have a measurable impact (especially on boot time) and the code isn't obscured too much then we're definitely open to those sorts of contributions 🙂