@neosilky 2019-05-28
> If they have a measurable impact (especially on boot time) and the code isn't obscured too much then we're definitely open to those sorts of contributions slightly_smiling_face

Ok cool, will see what I have :+1: