@rishabhgrg 2019-05-28
@kevinansfield Raised an issue [here](https://github.com/TryGhost/Members/issues/27) to track update for members keypair generation, overall agree its a good idea to move keypair away from boot if possible at best, or only on first boot at worst. :)