@allouis 2019-06-17
Hey @neosilky Only just got round to catching up with this.

@kevinansfield @rishabhgrg This looks like a sane and solid change for this - I think we should merge it regardless of updates to keypair generation being skipped.

@rishabhgrg Are you able to move that issue to this repo as the changes concern files in here :+1: