@allouis 2019-06-22
Yo @neosilky thanks again for this PR (and the research/stats behind it!)

My laptop must be a bit slower than yours, because my boot time went from ~4.5 seconds to ~2.2 seconds! This has made my life so much better :sweat_smile: :ghost: