@neosilky 2019-06-22
> Yo @neosilky thanks again for this PR (and the research/stats behind it!)
> 
> My laptop must be a bit slower than yours, because my boot time went from ~4.5 seconds to ~2.2 seconds! This has made my life so much better sweat_smile ghost

Haha - excellent! Would love to hear if it made much of a difference for your Pro platform.

Need to get around to sending in my other changes... It's a fairly minimal speedup for me, but now I know it's even more beneficial on other machines, they should still be useful.